centralize validate code for levels and adventures
