+++ tests/tests_z_defaults.py
@@ -1,9 +1,9 @@
 import os
-import re
 from website.yaml_file import YamlFile
 import utils
 import hedy
 import unittest
+from Tester import HedyTester
 
 # file is called tests_z_ so they are executed last
 # because programs are more of a priority than level defaults, which change less and take longer to run
@@ -12,36 +12,12 @@ import unittest
 os.chdir(os.path.join (os.getcwd (), __file__.replace (os.path.basename (__file__), '')))
 
 path = '../coursedata/level-defaults'
-files = [f for f in os.listdir(path) if os.path.isfile(os.path.join(path, f)) and f.endswith ('.yaml')]
-
-def check_code(filename, level, field_name, code):
-    # We ignore empty code snippets
-    if len(code) == 0:
-        return True
-    try:
-        hedy.transpile(code, int(level))
-    except Exception as E:
-        if len(E.args) == 0:
-            filename_shorter = filename.split("/")[3]
-            language = filename_shorter.split(".")[0]
-            error = f'{language}: level #{level} - {field_name}. Error: {E.args}'
-            # We print the error for readability, since otherwise they get accumulated on a long list
-            print(error)
-            return error
-        else:
-            if E.args[0] != 'Has Blanks':  # code with blanks is ok!
-                filename_shorter = filename.split("/")[3]
-                language = filename_shorter.split(".")[0]
-                error = f'{language}: level #{level} - {field_name}. Error: {E.args[0]}'
-                # We print the error for readability, since otherwise they get accumulated on a long list
-                print(error)
-                return error
-    return True
+files =[f for f in os.listdir(path) if os.path.isfile(os.path.join(path, f)) and f.endswith ('.yaml')]
 
 class TestsLevelDefaultsPrograms(unittest.TestCase):
 
     def test_level_defaults_snippets(self):
-        level_default_fails = []
+        level_default_fails =[]
 
         for file in files:
             file = os.path.join (path, file)
@@ -49,23 +25,24 @@ class TestsLevelDefaultsPrograms(unittest.TestCase):
 
             for level in yaml:
                 # start_code
-                result = check_code(file, level, 'start_code', yaml [level] ['start_code'])
+                result = HedyTester.validate_Hedy_code(file, level, 'start_code', yaml[level]['start_code'], 'level_defaults')
                 if result != True:
                     level_default_fails.append(result)
                 # commands.k.demo_code
                 for k, command in enumerate(yaml[level]['commands']):
+                    # todo: has a name now (again?)
                     command_text_short = command['explanation'][0:10]
-                    result = check_code(file, level, 'command ' + command_text_short + ' demo_code', command['demo_code'])
+                    result = HedyTester.validate_Hedy_code(file, level, 'command ' + command_text_short + ' demo_code', command['demo_code'], 'level_defaults')
                     if result != True:
                         level_default_fails.append(result)
 
                 # code snippets inside intro_text
                 code_snippet_counter = 0
-                for tag in utils.markdown_to_html_tags (yaml [level] ['intro_text']):
-                    if tag.name != 'pre' or not tag.contents [0]:
+                for tag in utils.markdown_to_html_tags (yaml[level]['intro_text']):
+                    if tag.name != 'pre' or not tag.contents[0]:
                         continue
                     code_snippet_counter += 1
-                    result = check_code(file, level, 'intro_text snippet #' + str (code_snippet_counter), tag.contents [0].contents [0])
+                    result = HedyTester.validate_Hedy_code(file, level, 'intro_text snippet #' + str (code_snippet_counter), tag.contents[0].contents[0], 'level_defaults')
                     if result != True:
                         level_default_fails.append(result)
